requirements.txt
djangorestframework
certifi==2016.2.28
Django==2.0
mysqlclient==1.3.12
pytz==2017.3
